fix to api call to view in app
