+++ app/src/main/java/com/example/kotlin/newapp/utils/Constants.kt
@@ -2,4 +2,5 @@ package com.example.kotlin.newapp.utils
 
 object Constants {
     const val BASE_URL = "https://api.themoviedb.org/3/movie/"
+    const val language = "en-US"
 }
